+++ FastCoder/FastCoder.h
@@ -1,7 +1,7 @@
 //
 //  FastCoding.h
 //
-//  Version 1.1
+//  Version 2.0 beta
 //
 //  Created by Nick Lockwood on 09/12/2013.
 //  Copyright (c) 2013 Charcoal Design
@@ -32,6 +32,14 @@
 
 
 #import <Foundation/Foundation.h>
+
+
+@interface NSObject (FastCoding)
+
++ (NSArray *)fastCodingKeys;
+- (id)awakeAfterFastCoding;
+
+@end
 
 
 @interface FastCoder : NSObject